Sharding-JDBC大规模数据示例重命名为Sharding-JDBC Benchmark示例

+++ demo-mysql-n-mariadb/demo-sharding-jdbc-benchmark/src/main/java/com/future/demo/bean/Order.java
@@ -7,7 +7,8 @@ import java.time.LocalDateTime;
 
 @Data
 public class Order {
-    private BigDecimal id;
+//    private BigDecimal id;
+    private Long id;
     private LocalDateTime createTime;
     private Long userId;
     private Long merchantId;
